+++ tests/views/mail/plain.php
@@ -1,2 +1,9 @@
+<?php
+/**
+ * @var $this \yii1tech\mailer\View
+ * @var $_message \yii1tech\mailer\TemplatedEmail|null
+ */
+?>
 Test plain mail template
 Name = <?php echo $name; ?>
+Subject = <?php echo isset($_message) ? $_message->getSubject() : ''; ?>