+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Yii1 Symfony Mailer extension
 =============================
 
+1.0.1 Under Development
+-----------------------
+
+- Enh: Added passing email message instance via `$_message` context variable to the template (klimov-paul)
+- Enh: Added passing template context variables to the layout template (klimov-paul)
+
+
 1.0.0, May 10, 2024
 -------------------